add click count
